refactor: HTTP testoperator
