--- crates/test-framework/src/spicetest/consistency/mod.rs
@@ -1,230 +1,0 @@
-/*
-Copyright 2024-2025 The Spice.ai OSS Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-use anyhow::anyhow;
-use std::{
-    sync::Arc,
-    time::{Duration, Instant},
-};
-
-use async_openai::{
-    config::OpenAIConfig,
-    types::{
-        ChatCompletionRequestMessage, ChatCompletionRequestUserMessageArgs,
-        CreateChatCompletionRequest, CreateChatCompletionRequestArgs, CreateEmbeddingRequest,
-        EncodingFormat,
-    },
-    Client as OpenAIClient,
-};
-use reqwest::Client;
-use worker::{ConsistencyWorker, ConsistencyWorkerResult, WorkerHandle};
-
-use crate::{
-    metrics::{NoExtendedMetrics, QueryMetric},
-    spiced::SpicedInstance,
-};
-
-mod worker;
-
-#[derive(Clone)]
-pub struct ConsistencyConfig {
-    /// The total duration of the test.
-    pub duration: Duration,
-
-    /// The number of buckets to divide the test duration into.
-    pub buckets: usize,
-
-    /// The number of individial HTTP clients to make requests in parallel.
-    pub concurrency: usize,
-
-    /// The payloads to send to the component, specifically to be used in [`HttpConsistencyComponent::send_request`].
-    pub payloads: Vec<Arc<str>>,
-
-    /// The HTTP component, within the Spiced instance, to test.
-    pub component: HttpConsistencyComponent,
-}
-
-/// A component within the Spiced instance to test for consistency.
-///
-/// This component must be accessible over HTTP.
-#[derive(Clone)]
-pub enum HttpConsistencyComponent {
-    Model { model: String, api_base: String },
-    Embedding { embedding: String, api_base: String },
-}
-
-impl HttpConsistencyComponent {
-    fn api_base(&self) -> String {
-        match self {
-            HttpConsistencyComponent::Model { api_base, .. } => api_base.clone(),
-            HttpConsistencyComponent::Embedding { api_base, .. } => api_base.clone(),
-        }
-    }
-
-    /// Sends a request to the component and returns the duration of the request.
-    /// Payload may be the entire HTTP request body, or a portion of it (dependent of the component).
-    pub async fn send_request(&self, client: &Client, payload: &str) -> anyhow::Result<Duration> {
-        let c = OpenAIClient::with_config(OpenAIConfig::default().with_api_base(self.api_base()))
-            .with_http_client(client.clone())
-            .clone();
-
-        let start_time = Instant::now();
-        match self {
-            HttpConsistencyComponent::Model { model, .. } => {
-                let req: CreateChatCompletionRequest =
-                    match serde_json::from_str::<CreateChatCompletionRequest>(payload) {
-                        Ok(mut req) => {
-                            // Ensure the model is overriden.
-                            req.model = model.clone();
-                            req
-                        }
-                        Err(_) => CreateChatCompletionRequestArgs::default()
-                            .model(model.clone())
-                            .messages(vec![ChatCompletionRequestMessage::User(
-                                ChatCompletionRequestUserMessageArgs::default()
-                                    .content(payload.to_string())
-                                    .build()
-                                    .map_err(|e| {
-                                        anyhow!("failed to build user message. Error: {e:?}")
-                                    })?,
-                            )])
-                            .build()
-                            .map_err(|e| anyhow!("Failed to build model request. Error: {e:?}"))?,
-                    };
-                let _ = c.chat().create(req).await?;
-            }
-            HttpConsistencyComponent::Embedding { embedding, .. } => {
-                let req: CreateEmbeddingRequest =
-                    match serde_json::from_str::<CreateEmbeddingRequest>(payload) {
-                        Ok(mut req) => {
-                            // Ensure the model is overriden.
-                            req.model = embedding.clone();
-                            req
-                        }
-                        Err(_) => CreateEmbeddingRequest {
-                            model: embedding.clone(),
-                            input: async_openai::types::EmbeddingInput::String(payload.to_string()),
-                            encoding_format: Some(EncodingFormat::Float),
-                            user: None,
-                            dimensions: None,
-                        },
-                    };
-                c.embeddings().create(req).await?;
-            }
-        }
-        Ok(start_time.elapsed())
-    }
-}
-
-pub enum ConsistencyState {
-    NotStarted,
-    Running { worker_handles: Vec<WorkerHandle> },
-    Completed { result: ConsistencyWorkerResult },
-}
-
-pub struct ConsistencySpiceTest {
-    start_time: Option<Instant>,
-    spiced_instance: SpicedInstance,
-    config: ConsistencyConfig,
-    state: ConsistencyState,
-}
-
-impl ConsistencySpiceTest {
-    #[must_use]
-    pub fn new(spiced_instance: SpicedInstance, config: ConsistencyConfig) -> Self {
-        Self {
-            start_time: None,
-            spiced_instance,
-            config,
-            state: ConsistencyState::NotStarted,
-        }
-    }
-
-    pub async fn start(self) -> anyhow::Result<ConsistencySpiceTest> {
-        if !matches!(self.state, ConsistencyState::NotStarted) {
-            return Err(anyhow::anyhow!("Test already started"));
-        };
-
-        let client = self.spiced_instance.http_client()?;
-
-        let worker_handles = (0..self.config.concurrency)
-            .map(|id| {
-                let worker = ConsistencyWorker::new(id, self.config.clone(), client.clone());
-                worker.start()
-            })
-            .collect();
-
-        Ok(ConsistencySpiceTest {
-            start_time: Some(Instant::now()),
-            spiced_instance: self.spiced_instance,
-            config: self.config,
-            state: ConsistencyState::Running { worker_handles },
-        })
-    }
-
-    pub async fn wait(self) -> anyhow::Result<ConsistencySpiceTest> {
-        match self.state {
-            ConsistencyState::Running { worker_handles } => {
-                let mut error_count = 0;
-
-                let mut durations: Vec<Vec<Duration>> = vec![vec![]; self.config.buckets];
-
-                for worker_handle in worker_handles {
-                    match worker_handle.await {
-                        Ok(worker_result) => {
-                            for (i, minute) in worker_result.durations.iter().enumerate() {
-                                durations[i].extend(minute);
-                            }
-                            error_count += worker_result.error_count;
-                        }
-                        Err(_) => {
-                            return Err(anyhow::anyhow!("Worker failed"));
-                        }
-                    }
-                }
-
-                Ok(ConsistencySpiceTest {
-                    start_time: self.start_time,
-                    spiced_instance: self.spiced_instance,
-                    config: self.config,
-                    state: ConsistencyState::Completed {
-                        result: ConsistencyWorkerResult {
-                            durations,
-                            error_count,
-                        },
-                    },
-                })
-            }
-            ConsistencyState::NotStarted => Err(anyhow::anyhow!("Test not started")),
-            ConsistencyState::Completed { .. } => Err(anyhow::anyhow!("Test already completed")),
-        }
-    }
-
-    pub fn get_result(&self) -> anyhow::Result<Vec<QueryMetric<NoExtendedMetrics>>> {
-        match self.state {
-            ConsistencyState::Completed { ref result, .. } => result
-                .durations
-                .iter()
-                .enumerate()
-                .map(|(i, durations)| {
-                    QueryMetric::new_from_durations(format!("Minute {i}").as_str(), durations)
-                })
-                .collect(),
-            ConsistencyState::NotStarted => Err(anyhow::anyhow!("Test not started")),
-            ConsistencyState::Running { .. } => Err(anyhow::anyhow!("Test still running")),
-        }
-    }
-}
