+++ crates/test-framework/src/spicetest/http/mod.rs
@@ -1,0 +1,170 @@
+/*
+Copyright 2024-2025 The Spice.ai OSS Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+     https://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+use super::{SpiceTest, TestCompleted, TestNotStarted, TestState};
+use crate::metrics::{MetricCollector, NoExtendedMetrics, QueryMetric};
+use anyhow::Result;
+use component::HttpComponent;
+use std::{
+    sync::Arc,
+    time::{Duration, SystemTime},
+};
+use worker::{ConsistencyWorker, ConsistencyWorkerResult, WorkerHandle};
+
+pub mod component;
+mod worker;
+
+#[derive(Clone)]
+pub struct HttpConfig {
+    /// The total duration of the test.
+    pub duration: Duration,
+
+    /// The number of buckets to divide the test duration into.
+    pub buckets: usize,
+
+    /// The number of individial HTTP clients to make requests in parallel.
+    pub concurrency: usize,
+
+    /// The payloads to send to the component, specifically to be used in [`HttpComponent::send_request`].
+    pub payloads: Vec<Arc<str>>,
+
+    /// The HTTP component, within the Spiced instance, to test.
+    pub component: HttpComponent,
+}
+
+pub struct NotStarted {
+    config: HttpConfig,
+}
+
+impl NotStarted {
+    #[must_use]
+    pub fn new(config: HttpConfig) -> Self {
+        Self { config }
+    }
+}
+
+pub struct Running {
+    config: HttpConfig,
+    worker_handles: Vec<WorkerHandle>,
+}
+
+pub struct Completed {
+    result: ConsistencyWorkerResult,
+    end_time: SystemTime,
+}
+
+impl TestState for NotStarted {}
+impl TestState for Running {}
+impl TestState for Completed {}
+impl TestNotStarted for NotStarted {}
+impl TestCompleted for Completed {
+    fn end_time(&self) -> SystemTime {
+        self.end_time
+    }
+}
+
+impl SpiceTest<NotStarted> {
+    pub fn start(self) -> Result<SpiceTest<Running>> {
+        let client = self.spiced_instance.http_client()?;
+
+        let worker_handles = (0..self.state.config.concurrency)
+            .map(|id| {
+                let worker = ConsistencyWorker::new(id, self.state.config.clone(), client.clone());
+                worker.start()
+            })
+            .collect::<Vec<_>>();
+
+        Ok(SpiceTest {
+            name: self.name,
+            start_time: self.start_time,
+            spiced_instance: self.spiced_instance,
+            use_progress_bars: self.use_progress_bars,
+            state: Running {
+                worker_handles,
+                config: self.state.config,
+            },
+        })
+    }
+}
+
+impl SpiceTest<Running> {
+    pub async fn wait(self) -> Result<SpiceTest<Completed>> {
+        let mut error_count = 0;
+
+        let mut durations: Vec<Vec<Duration>> = vec![vec![]; self.state.config.buckets];
+
+        for worker_handle in self.state.worker_handles {
+            match worker_handle.await? {
+                Ok(worker_result) => {
+                    for (i, minute) in worker_result.durations.iter().enumerate() {
+                        durations[i].extend(minute);
+                    }
+                    error_count += worker_result.error_count;
+                }
+                Err(_) => {
+                    return Err(anyhow::anyhow!("Worker failed"));
+                }
+            }
+        }
+
+        Ok(SpiceTest {
+            name: self.name,
+            start_time: self.start_time,
+            spiced_instance: self.spiced_instance,
+            use_progress_bars: self.use_progress_bars,
+            state: Completed {
+                result: ConsistencyWorkerResult {
+                    durations,
+                    error_count,
+                },
+                end_time: SystemTime::now(),
+            },
+        })
+    }
+}
+
+impl SpiceTest<Completed> {
+    #[must_use]
+    pub fn get_durations(&self) -> &Vec<Vec<Duration>> {
+        &self.state.result.durations
+    }
+}
+
+impl MetricCollector<NoExtendedMetrics> for SpiceTest<Completed> {
+    fn start_time(&self) -> SystemTime {
+        self.start_time
+    }
+
+    fn end_time(&self) -> SystemTime {
+        self.state.end_time()
+    }
+
+    fn name(&self) -> String {
+        self.name.clone()
+    }
+
+    fn metrics(&self) -> Result<Vec<QueryMetric<NoExtendedMetrics>>> {
+        self.state
+            .result
+            .durations
+            .iter()
+            .enumerate()
+            .map(|(i, durations)| {
+                QueryMetric::new_from_durations(format!("Minute {i}").as_str(), durations)
+            })
+            .collect()
+    }
+}
