+++ tools/testoperator/src/tests/load/mod.rs
@@ -22,57 +22,76 @@ use test_framework::{
     metrics::{MetricCollector, StatisticsCollector},
     queries::{QueryOverrides, QuerySet},
     spiced::SpicedInstance,
-    spicetest::{EndCondition, SpiceTest},
+    spicetest::{
+        datasets::{EndCondition, NotStarted},
+        SpiceTest,
+    },
+    TestType,
 };
 
+#[allow(clippy::too_many_lines)]
 pub(crate) async fn run(args: &DatasetTestArgs) -> anyhow::Result<()> {
+    if args.common.concurrency < 2 {
+        return Err(anyhow::anyhow!(
+            "Concurrency should be greater than 1 for a load test"
+        ));
+    }
+
     let query_set = QuerySet::from(args.query_set.clone());
     let query_overrides = args.query_overrides.clone().map(QueryOverrides::from);
     let queries = query_set.get_queries(query_overrides);
 
-    let (app, start_request) = get_app_and_start_request(args)?;
+    let (app, start_request) = get_app_and_start_request(&args.common)?;
     let mut spiced_instance = SpicedInstance::start(start_request).await?;
 
     spiced_instance
-        .wait_for_ready(Duration::from_secs(args.ready_wait.unwrap_or(30) as u64))
+        .wait_for_ready(Duration::from_secs(args.common.ready_wait))
         .await?;
 
-    let test_duration = Duration::from_secs(args.duration.unwrap_or(60).try_into()?);
+    let test_duration = Duration::from_secs(args.common.duration);
     let test_hours = (test_duration.as_secs() / 60 / 60).min(1);
 
     // baseline run
     println!("Running baseline throughput test");
-    let baseline_test = SpiceTest::new(app.name.clone(), spiced_instance)
-        .with_query_set(queries.clone())
-        .with_parallel_count(args.concurrency.unwrap_or(8))
-        .with_end_condition(EndCondition::QuerySetCompleted(test_hours.try_into()?))
-        .with_progress_bars(!args.disable_progress_bars)
-        .start()
-        .await?;
+    let baseline_test = SpiceTest::new(
+        app.name.clone(),
+        spiced_instance,
+        NotStarted::new()
+            .with_parallel_count(args.common.concurrency)
+            .with_query_set(queries.clone())
+            .with_end_condition(EndCondition::QuerySetCompleted(test_hours.try_into()?)),
+    )
+    .with_progress_bars(!args.common.disable_progress_bars)
+    .start()
+    .await?;
 
     let test = baseline_test.wait().await?;
     let baseline_percentiles = test
         .get_query_durations()
         .statistical_set()?
         .percentile(0.99)?;
-    let baseline_metrics = test.collect()?;
+    let baseline_metrics = test.collect(TestType::Load)?;
     let spiced_instance = test.end();
 
     // load test
     println!("Running load test");
-    let throughput_test = SpiceTest::new(app.name.clone(), spiced_instance)
-        .with_query_set(queries.clone())
-        .with_parallel_count(args.concurrency.unwrap_or(8))
-        .with_end_condition(EndCondition::Duration(Duration::from_secs(
-            args.duration.unwrap_or(60).try_into()?,
-        )))
-        .with_progress_bars(!args.disable_progress_bars)
-        .start()
-        .await?;
+    let throughput_test = SpiceTest::<NotStarted>::new(
+        app.name.clone(),
+        spiced_instance,
+        NotStarted::new()
+            .with_parallel_count(args.common.concurrency)
+            .with_query_set(queries.clone())
+            .with_end_condition(EndCondition::Duration(Duration::from_secs(
+                args.common.duration,
+            ))),
+    )
+    .with_progress_bars(!args.common.disable_progress_bars)
+    .start()
+    .await?;
 
     let test = throughput_test.wait().await?;
     let test_durations = test.get_query_durations().statistical_set()?;
-    let metrics = test.collect()?;
+    let metrics = test.collect(TestType::Load)?;
     let mut spiced_instance = test.end();
 
     println!("Baseline metrics:");
